place optim state_dict loading inside `load_checkpoint` (#839)
## Description

To bring all the checkpoint loading operations into one single API, I
place the optim load_state_dict code inside load_checkpoint.

## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)
- [x] This change requires a documentation update

## Checklist:

- [x] Have you added tests that prove your fix is effective or that this
feature works?
- [x] Has code been commented, particularly in hard-to-understand areas?
- [x] Have you made corresponding changes to the documentation?

---------
